pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "violation-detection-system"
version = "1.0.0"
description = "基于多模态AI的违章建筑智能检测与管理系统"
authors = [
    {name = "Developer", email = "developer@example.com"},
]
requires-python = ">=3.9"
dependencies = [
    "ultralytics>=8.0.196",
    "torch>=1.13.0",
    "torchvision>=0.14.0",
    "opencv-python>=4.8.1",
    "Pillow>=9.0.0",
    "fastapi>=0.103.1",
    "uvicorn[standard]>=0.23.2",
    "python-multipart>=0.0.6",
    "numpy>=1.24.0",
    "pandas>=1.5.0",
    "asyncpg>=0.28.0",
    "redis>=4.0.0",
    "sqlalchemy[asyncio]>=2.0.21",
    "alembic>=1.12.0",
    "pydantic>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.1",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.2",
    "pytest-asyncio>=0.21.1",
    "black>=23.9.1",
    "flake8>=6.1.0",
    "mypy>=1.5.1",
    "isort>=5.12.0",
    "pre-commit>=3.4.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"